Rename HttpBodyImpl to BufferedBody
