+++ src/core/http-body/buffered-body.ts
@@ -1,7 +1,7 @@
-import { HttpBody } from "./http";
-import { TheTextDecoder } from "./env";
+import { HttpBody } from "../http";
+import { TheTextDecoder } from "../env";
 
-export class HttpBodyImpl implements HttpBody {
+export class BufferedBody implements HttpBody {
   constructor(private readonly it: AsyncIterable<Uint8Array>) {}
 
   async *[Symbol.asyncIterator]() {
@@ -26,7 +26,7 @@ export class HttpBodyImpl implements HttpBody {
       yield new Uint8Array(contentArr);
     }
 
-    return new HttpBodyImpl(gen());
+    return new BufferedBody(gen());
   }
 }
 
@@ -38,7 +38,7 @@ export async function asJson<T>(body: HttpBody): Promise<T> {
 
 export function jsonBody(obj: any) {
   const objStr = JSON.stringify(obj);
-  return HttpBodyImpl.fromString(objStr);
+  return BufferedBody.fromString(objStr);
 }
 
 // TODO: Move static function `fromString` out of Body and create StringBody and JsonBody classes
