+++ src/core/mod.ts
@@ -2,6 +2,6 @@ export * from "./env";
 export * from "./http-status";
 export * from "./http";
 export * from "./http4ts";
-export * from "./HttpBodyImpl";
+export * from "./http-body/buffered-body";
 export * from "./HttpRequestImpl";
 export * from "./routing/routes";
