+++ src/core/routing/__tests__/routes.spec.ts
@@ -1,7 +1,7 @@
 import { HttpHeaders, HttpRequest, HttpResponse, HttpMethod } from "../../http";
 import { routes, get, notFound, RoutedHttpRequest, post, all } from "../routes";
 import { HttpRequestImpl } from "../../HttpRequestImpl";
-import { HttpBodyImpl } from "../../HttpBodyImpl";
+import { BufferedBody } from "../../http-body/buffered-body";
 
 function resp(
   status = 200,
@@ -10,13 +10,13 @@ function resp(
 ): HttpResponse {
   return {
     status,
-    body: HttpBodyImpl.fromString(body),
+    body: BufferedBody.fromString(body),
     headers
   };
 }
 
 function request(url: string, body: string, method: HttpMethod): HttpRequest {
-  return new HttpRequestImpl(url, HttpBodyImpl.fromString(body), method);
+  return new HttpRequestImpl(url, BufferedBody.fromString(body), method);
 }
 
 async function assertResponseEquality(
