+++ src/examples/simple-node-server.ts
@@ -5,13 +5,13 @@ import {
   HttpResponse,
   HttpStatus,
   toNodeRequestListener,
-  HttpBodyImpl
+  BufferedBody
 } from "../node";
 
 async function handler(req: HttpRequest): Promise<HttpResponse> {
   await req.body.asString("UTF-8");
   return {
-    body: HttpBodyImpl.fromString("This is some string"),
+    body: BufferedBody.fromString("This is some string"),
     headers: {},
     status: HttpStatus.OK
   };
